backend/requirements.txt
fastapi==0.115.0
httpx==0.27.0
numpy==1.26.4
Pillow==10.4.0
pydantic==2.9.2
pypdf==5.0.0
pytesseract==0.3.13
python-dotenv==1.0.1
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
torch==2.3.1
transformers==4.44.2
uvicorn[standard]==0.30.6
PyYAML==6.0.1
slowapi==0.1.9
pytest==8.3.3
pytest-asyncio==0.24.0
pytest-cov==5.0.0
factory_boy==3.3.0
